rename FleetSummary component to ResourcesSummary
