+++ components/fleet/ResourcesSummary.vue
